refactor: inventory page
