+++ app/src/main/java/bangkit/kiki/foodwisemobile/ui/inventory/component/InventoryItem.kt
@@ -1,0 +1,89 @@
+package bangkit.kiki.foodwisemobile.ui.inventory.component
+
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.IntrinsicSize
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.layout.width
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material.Card
+import androidx.compose.material3.Icon
+import androidx.compose.material3.IconButton
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import bangkit.kiki.foodwisemobile.R
+
+@Composable
+fun InventoryItem(
+    id: Int,
+    name: String,
+    expiryDate: String,
+    quantity: Number,
+    measure: String,
+    showIconButton: Boolean,
+    onDeleteClick: (Int) -> Unit
+) {
+    Card(
+        modifier = Modifier
+            .fillMaxWidth()
+            .padding(horizontal = 16.dp, vertical = 5.dp),
+        shape = RoundedCornerShape(8.dp),
+        elevation = 8.dp,
+        backgroundColor = Color.White
+    ) {
+        Row(
+            modifier = Modifier
+                .fillMaxWidth()
+                .height(IntrinsicSize.Min),
+            verticalAlignment = Alignment.CenterVertically,
+        ) {
+            Card(
+                modifier = Modifier.size(90.dp),
+                shape = RoundedCornerShape(8.dp),
+                elevation = 8.dp,
+                backgroundColor = Color.White
+            ) {
+                Box(
+                    contentAlignment = Alignment.Center
+                ) {
+                    Text(
+                        text = "x$quantity",
+                        fontSize = 30.sp,
+                        fontWeight = FontWeight.SemiBold
+                    )
+                }
+            }
+            Spacer(modifier = Modifier.width(16.dp))
+            Column(
+                modifier = Modifier.weight(1f)
+            ) {
+                Text(name, fontSize = 20.sp, fontWeight = FontWeight.Bold)
+                Text(measure, fontSize = 16.sp)
+                Text("Expired at $expiryDate", fontSize = 14.sp, color = Color.Gray)
+            }
+            if (showIconButton) {
+                Box(modifier = Modifier.padding(end = 8.dp)) {
+                    IconButton(onClick = { onDeleteClick(id) }) {
+                        Icon(
+                            painter = painterResource(id = R.drawable.ic_eat),
+                            contentDescription = "Delete",
+                            tint = Color.Red
+                        )
+                    }
+                }
+            }
+        }
+    }
+}
