+++ app/src/main/java/bangkit/kiki/foodwisemobile/ui/main/MainActivity.kt
@@ -69,6 +69,6 @@ fun DefaultPreview() {
 @Composable
 fun ProgressPreview() {
     FoodwiseMobileTheme {
-        InventoryStatisticsSection()
+        ExpiringSection()
     }
 }
